requirements.txt
alabaster==0.7.12
asgiref==3.2.7
Babel==2.8.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
commonmark==0.9.1
Django==3.0.6
django-templated-mail==1.1.1
djangorestframework==3.11.0
docutils==0.16
google-auth==1.14.2
google-auth-oauthlib==0.4.1
gspread==3.6.0
gspread-formatting==0.2.4
httplib2==0.17.3
idna==2.9
imagesize==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pytz==2020.1
recommonmark==0.6.0
requests==2.23.0
requests-oauthlib==1.3.0
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.4.0
rsa==4.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==1.7.5
sphinxcontrib-websupport==1.2.2
sqlparse==0.3.1
urllib3==1.25.9
cryptography~=2.8
django-cors-headers~=3.1.1
drf-jwt~=1.13.3
pyjwt~=1.7.1